Handle additional server errors in socket.{close,accept}
When closing the server socket, rescue IOError and exit.

Handle ThreadError and TypeError when accept fails. This shouldn't
happen, and probably signals a bug in the Ruby implementation on
the platform, but I've seen it. Exit the accept loop if it occurs.

While here, add remove_server_socket helper method, which doesn't
raise if the socket doesn't exist, and use it when removing the
socket.